finish crawling uci and aws

+++ src/main.py
@@ -6,6 +6,7 @@ def main():
     mode = 'local'
     is_update = True  # set to true for first run when vdb was renewed or new documents inserted in mongodb
     config_path = 'config.yml'
+    setup_milvus = False
     with open(config_path, 'r') as file:
         config = yaml.safe_load(file)
     database = coldata.mongodb.MongoDB(mode=mode, **config['mongodb'])
@@ -22,22 +23,23 @@ def main():
     aws = coldata.crawler.AWS(database, **config['crawler'])
     aws.crawl()
     aws.upload()
-    '''
-    pwc = coldata.crawler.PapersWithCode(database, **config['crawler'])
-    pwc.crawl()
-    pwc.upload()
-    '''
-    vdb = coldata.vdb.VDB(**config['vdb']['milvus'], **config['vdb']['text'], **config['vdb']['model'])
-    if is_update or config['vdb']['milvus']['renew']:
-        vdb.update(database)
-    result = vdb.search(database, ['Satellite Computed Bathymetry Assessment-SCuBA'])
-    for i in range(len(result)):
-        for index in result[i]:
-            print(result[i][index]['distance'])
-            print(result[i][index])
 
-    if is_update:
-        vdb.drop()
+    # pwc = coldata.crawler.PapersWithCode(database, **config['crawler'])
+    # pwc.crawl()
+    # pwc.upload()
+
+    if setup_milvus:
+        vdb = coldata.vdb.VDB(**config['vdb']['milvus'], **config['vdb']['text'], **config['vdb']['model'])
+        if is_update or config['vdb']['milvus']['renew']:
+            vdb.update(database)
+        result = vdb.search(database, ['Satellite Computed Bathymetry Assessment-SCuBA'])
+        for i in range(len(result)):
+            for index in result[i]:
+                print(result[i][index]['distance'])
+                print(result[i][index])
+
+        if is_update:
+            vdb.drop()
     return
